order DATA_COLUMS automatically, Measurable, analyse function
Added DATA_COLUMS list to Measurable class which keeps track of the
order in which they were created.
Added ‘analyse’ function for on-the-fly analysis for live plotting.
Added the functionality to manually set a Measurable parameter, for
e.g. in-measurement data analysis or keeping track of iterators.
Updated the examples script.py and script2.py to illustrate above
changes.