Created new joined_validator, which allows for multiple validators to be connected together.
The specific validator needs to be created before use as it returns a
validator to be further used in Instrument IO validation.

The values parameter of any Instrument IO must be a list of values to be
accepted for each respective validator in the chain.

A test case was added for the extra functionality.