Enable Read/Write delay on Prologix adpaters.
Prologix will return an empty string from an ask request if the instrument is not able to
process the write command fast enough before the read command is issued.

This commit allows for a simple keyword parameter, `rw_delay` to the adapter instantiation,
which is a value in seconds to wait between the read and write of any `ask` call.