Fix ScientificInput GUI min/max parameters not set
Input parent constructor is called after min/max are set, causing
min/max to be reset to their defaults. Moved the Input constructor
call up and moved the min/max set to set_parameter(), as it is called
by Input.__init__() and min/max would have to be called on later calls
to set_parameter().