class ManagedWindow and ManagedImageWindow are now related with ManagedImageWindow subclass of ManagedWindow. ManagedWindows is subclass of ManagedWindowBase.
The idea also is to allow the user to define a generic list of widget to populate the MainWindow QTabWidget. Each widget should support some specific method.
A new base class ManagedWindowBase is now created and then the old ManagedWindow is now a subclass of ManagedWindowBase with two "Tab" widget as LogWidget and PlotWidget.
The old ManagedImageWindow is now a subclass of ManagedWindow which adds ImageWidget to the list.
In general, the user can subclass ManagedWindowBase to define its own list of Widget representing different views of the experiment. Some possible useful widget are: documentation of the experiment, tabular view of experiment data, etc.

The changes are preliminary and intended to trigger feedback especially from the people more experienced with the implementation.