add additional properties to AndeenHagerling brides
measurement was extended to also return the used voltage. The highest allowed voltage the bridge is allowed to use can not be set using vhighest.

The frequency can only be set for the AH2700A. Its fixed to 1kHz for the AH2500A.

Note: AH2500A functions are tested
Note: AH2700A still needs testing