keep a reference to experiment in analysis (#6)
This PR adds a reference to the experiment that invoked the analysis on the Analysis object so that when the Analysis.finished method is called the browser widget can be told to redraw the analyzed curve. Why? Some data is calculated and added to the data columns after the fact.