symetry in setting/getting properties
Some properties use integer to set their actual floating point values. The driver gives the possibility to get the integer or the float directly but you could only set it using the integer representation.

That creates an asymetry not working well with pymeasure Control object. I therefore changed the get command to the one using the integer (without the dot)