add MKS 974B vacuum pressure transducer (#1034)
* add MKS974 pressure transducer Instrument

* create an abstract MKSInstrument to avoid code duplication for the two
MKS instruments.
* add documentation
* add unit tests

* fix docstrings line length

* fix whitespace

* more whitespace fixes

* fixes after testing all options with the hardware

also add an id() implementation.
Note that this rapid succession of property queries in the id function might fail with some cheap USB-Serial converters. But the MKS hardware can handle it.

* Apply suggestions from code review

Co-authored-by: Benedikt Burger <67148916+BenediktBurger@users.noreply.github.com>

* rename SetpointChannel to RelayChannel

at the same time I renamed the SetpointChannel.value to RelayChannel.setpoint

* harmonize use of mksinst Unit

use specific imports from standard library

Units for MKS937B and MKS974B are unfortunately a bit different (MBAR vs. mBAR and missing MICRON)

* harmonize use of RelayChannel

add new feature of RelayChannels to MKS937B

* update documentation

* avoid StrEnum which needs python>=3.11

* fix docstring reference and more StrEnum

* code formatting fixes

* backward compatible setting of unit property for MKS937B

Setting the unit is now backward compatible, but reading the unit still changed. This depends on the order of the values in the dictionary.

* use StrEnum with a fallback implementation

* simplify Unit implementation

* Indicate minimal version at which fallback code can be removed

Co-authored-by: Benedikt Burger <67148916+BenediktBurger@users.noreply.github.com>

* full backward compatible version. Enum only for writing

* use StrEnum instead of old string values

this is a breaking change with respect to the instrument in the published package version and there a note is added in the changelog

---------

Co-authored-by: Benedikt Burger <67148916+BenediktBurger@users.noreply.github.com>