Update SBIS26 instrument and tests
The instrument class SBIS26 and its corresponding tests have been updated, adding new functionality and enhancing error treatment. Now it's possible to move the instrument to a specific or relative position, and a method to send the instrument to its mechanical home was implemented. Also, improved error checking while setting a property, properly reading an acknowledgement after a set command, and dealing with potential errors.